refactor(utils): break package-level import cycles
`skfolio.utils` is the most-imported subpackage (209 references) yet three of
its modules reached upward into estimator-layer packages, forming three
package-level cycles:

  utils/_factor_tools.py  -> linear_model  <-> linear_model -> utils.validation
  utils/_factor_tools.py  -> preprocessing <-> preprocessing -> utils.stats
  utils/validation.py     -> containers    <-> containers -> utils.tools

These resolved at import time only because each cycle closed at submodule
granularity, which made them latent rather than harmless: any import later
added to a leaf module could turn a working import into an ImportError far
from the change that caused it.

Relocate the upward-reaching code to the layer that owns it, keeping
single-consumer helpers with their callers and shared helpers with the
package that owns their operation:

- `validate_asset_panel` (plus its two private helpers `_check_fields_exist`
  and `_bad_observation`) moves from `utils/validation.py` to `base.py`,
  next to `BaseAssetPanelTransformer`. It sets `asset_names_` / `n_assets_`
  on the estimator and checks them when `reset=False`, so it belongs with
  estimator support rather than with the container. It is added to
  `skfolio.base.__all__` and documented in `docs/api.rst`.
  `utils/validation.py` keeps `validate_cross_sectional_data`, which needs
  nothing from containers.

- `_cross_sectional_neutralize` moves to
  `linear_model/_cross_sectional/_utils.py` as `_cs_neutralize`, matching the
  existing `_cs_*` helper naming. It is a weighted cross-sectional regression
  shared by two workflows, so it belongs with the package that owns that
  operation.

- `_neutralize_scores` moves to a module-level private function at the bottom
  of `alpha/_base.py`, its only production caller.

- `_neutralize_exposures` moves to the bottom of
  `prior/_characteristics_factor_model.py`, alongside the existing helpers.
  Its standardization, in-place updates and processing order are unchanged.

- `utils/_factor_tools.py` keeps the factor/family name-resolution helpers,
  which import only numpy and typing. `_powered_market_cap` had no caller and
  no test, and is removed.

`preprocessing` stays independent of `linear_model`, and `skfolio.utils` now
imports only `_constants`, `exceptions` and `typing`. The remaining layering
is a DAG:

  typing/_constants/exceptions <- utils <- {containers, linear_model} <- ...

Backward compatibility: `skfolio.utils.validation.validate_asset_panel` was
explicitly exported through `__all__` and is used when writing custom
descriptors, so the old name keeps working until 2.0. A module-level
`__getattr__` resolves it through a deferred import from `skfolio.base`,
returns the canonical function object and emits a `FutureWarning`;
unrecognized attributes raise `AttributeError`. The deprecated name stays in
`__all__` so wildcard imports keep working too, consistent with the existing
deprecation in `utils/sorting.py`.

Tests move with the code: validator behaviour to `tests/test_base.py`,
legacy-import compatibility to `tests/test_utils/test_validation.py`,
score neutralization to `tests/test_alpha/test_base.py`, and the
exposure-neutralization test now imports from
`skfolio.prior._characteristics_factor_model`.

Refs #276
