+++ src/skfolio/alpha/_evaluation.py
@@ -25,6 +25,7 @@ from skfolio._constants import (
     _IDIO_VARIANCES,
     _REGRESSION_WEIGHTS,
 )
+from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel, AssetPanelView, Field3D
 from skfolio.model_selection._validation import _route_params
 from skfolio.typing import FloatArray, IntArray, StrArray
@@ -44,7 +45,6 @@ from skfolio.utils.tools import (
     _validate_positive_integer,
     _validate_positive_real,
 )
-from skfolio.utils.validation import validate_asset_panel
 
 __all__ = [
     "AlphaForecastComparison",
