+++ src/skfolio/alpha/_ew_sharpe_optimal_alpha.py
@@ -18,6 +18,7 @@ from skfolio._constants import (
 )
 from skfolio.alpha import ForecastUnit
 from skfolio.alpha._base import BaseAlpha, BaseAlphaDescriptorComposition
+from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor import BaseDescriptor
 from skfolio.typing import FloatArray
@@ -28,7 +29,6 @@ from skfolio.utils.tools import (
     _validate_positive_real,
     half_life_to_decay_factor,
 )
-from skfolio.utils.validation import validate_asset_panel
 
 _FITTED_ATTR = "alpha_"
