+++ src/skfolio/alpha/_predictor_alpha.py
@@ -24,6 +24,7 @@ from skfolio._constants import (
 )
 from skfolio.alpha import ForecastUnit
 from skfolio.alpha._base import BaseAlpha, BaseAlphaDescriptorComposition
+from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor import BaseDescriptor
 from skfolio.preprocessing import BaseCSTransformer, CSWinsorizer
@@ -35,7 +36,6 @@ from skfolio.utils.tools import (
     check_estimator,
     half_life_to_decay_factor,
 )
-from skfolio.utils.validation import validate_asset_panel
 
 _FITTED_ATTR = "alpha_"
