+++ src/skfolio/descriptor/_liquidity/_ew_amihud_illiquidity.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 
 import numpy as np
 
+from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor._base import BaseDescriptor
 from skfolio.typing import FloatArray
@@ -17,7 +18,6 @@ from skfolio.utils.tools import (
     _validate_positive_real,
     half_life_to_decay_factor,
 )
-from skfolio.utils.validation import validate_asset_panel
 
 _FITTED_ATTR = "illiquidity_"
