+++ src/skfolio/descriptor/_lottery_demand/_max_return.py
@@ -9,12 +9,12 @@ from __future__ import annotations
 import numpy as np
 from numpy.lib.stride_tricks import sliding_window_view
 
+from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor._base import BaseDescriptor
 from skfolio.typing import FloatArray
 from skfolio.utils._array_buffer import _update_buffer
 from skfolio.utils.tools import _is_integer_number
-from skfolio.utils.validation import validate_asset_panel
 
 _FITTED_ATTR = "max_return_"
