+++ src/skfolio/descriptor/_passthrough.py
@@ -6,10 +6,10 @@
 
 from __future__ import annotations
 
+from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor._base import BaseDescriptor
 from skfolio.typing import FloatArray, IntArray
-from skfolio.utils.validation import validate_asset_panel
 
 
 class Passthrough(BaseDescriptor, stateless=True):
