+++ src/skfolio/descriptor/_short_interest/_short_interest.py
@@ -8,11 +8,11 @@ from __future__ import annotations
 
 import numpy as np
 
+from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor._base import BaseDescriptor
 from skfolio.typing import FloatArray
 from skfolio.utils.stats import safe_divide
-from skfolio.utils.validation import validate_asset_panel
 
 
 class ShortInterest(BaseDescriptor, stateless=True):
