+++ src/skfolio/factor_exposure/_one_hot_categorical_factors.py
@@ -8,10 +8,10 @@ from __future__ import annotations
 
 import numpy as np
 
+from skfolio.base import validate_asset_panel
 from skfolio.containers import MISSING_CATEGORY_CODE, AssetPanel
 from skfolio.factor_exposure._base import BaseFactorExposure
 from skfolio.typing import FloatArray, ObjArray
-from skfolio.utils.validation import validate_asset_panel
 
 
 class OneHotCategoricalFactors(BaseFactorExposure, stateless=True):
