+++ src/skfolio/prior/_characteristics_factor_model.py
@@ -28,10 +28,11 @@ from skfolio._constants import (
     _RETURNS,
 )
 from skfolio.alpha import BaseAlpha
-from skfolio.base import BaseComposition
+from skfolio.base import BaseComposition, validate_asset_panel
 from skfolio.containers import AssetPanel, InactivePolicy
 from skfolio.factor_exposure import BaseFactorExposure, DerivedFactor
 from skfolio.linear_model import BaseCSLinearModel, CSLinearRegression
+from skfolio.linear_model._cross_sectional._utils import _cs_neutralize
 from skfolio.moments import (
     BaseCovariance,
     EWCovariance,
@@ -39,7 +40,7 @@ from skfolio.moments import (
     RegimeAdjustedEWCovariance,
 )
 from skfolio.moments.variance import BaseVariance, RegimeAdjustedEWVariance
-from skfolio.preprocessing import CSWinsorizer
+from skfolio.preprocessing import CSStandardScaler, CSWinsorizer
 from skfolio.prior._base import BasePrior
 from skfolio.prior._empirical import EmpiricalPrior
 from skfolio.prior._model import FactorModel, ReturnDistribution
@@ -47,9 +48,13 @@ from skfolio.prior._model._family_constraint_basis import (
     FamilyConstraintBasis,
     compute_family_constraint_basis,
 )
-from skfolio.typing import AnyArray, BoolArray, FloatArray, StrArray
+from skfolio.typing import AnyArray, BoolArray, FloatArray, ObjArray, StrArray
 from skfolio.utils._array_buffer import _ArrayBuffer, _update_buffer
-from skfolio.utils._factor_tools import _neutralize_exposures
+from skfolio.utils._factor_tools import (
+    _expand_factor_names,
+    _factor_name_maps,
+    _resolve_factor_name,
+)
 from skfolio.utils.equations import _validate_factor_names_and_families
 from skfolio.utils.stats import corr_to_cov, cov_nearest, cov_to_corr, safe_divide
 from skfolio.utils.tools import (
@@ -61,7 +66,6 @@ from skfolio.utils.tools import (
     call_asset_panel_transform,
     check_estimator,
 )
-from skfolio.utils.validation import validate_asset_panel
 
 _FITTED_ATTR = "factor_model_"
 
@@ -2792,7 +2796,7 @@ class CharacteristicsFactorModel(BasePrior, BaseComposition):
 
         # Exclude pairs with missing alpha (e.g. assets within the alpha estimator
         # warmup) or missing exposures by zeroing their regression weights, mirroring
-        # `_cross_sectional_neutralize`. The regression weights derive from the lagged
+        # `_cs_neutralize`. The regression weights derive from the lagged
         # exposures and returns, so they can be positive where the current alpha or
         # exposure is non-finite, which the regressor rejects.
         valid = np.isfinite(alpha)[np.newaxis, :] & np.all(
@@ -3090,3 +3094,75 @@ def _assemble_asset_return_scenarios(
         sample_weight = sample_weight[-n_scenarios:].copy()
 
     return asset_return_scenarios, sample_weight
+
+
+def _neutralize_exposures(
+    cs_regressor: BaseCSLinearModel,
+    neutralize_against: dict[str, list[str]],
+    exposures: FloatArray,
+    benchmark_weights: FloatArray,
+    factor_names: ObjArray,
+    factor_families: ObjArray,
+) -> None:
+    """Neutralize factor exposures against specified factors or families.
+
+    For each entry in `neutralize_against`, the key's exposures are regressed against
+    the target factors and replaced in-place by standardized residuals. Keys and targets
+    accept factor names or family names. Entries are processed in insertion order, so
+    later entries see exposures already modified by earlier entries.
+
+    Parameters
+    ----------
+    cs_regressor : BaseCSLinearModel
+        Cross-sectional linear regressor used for residualization.
+
+    neutralize_against : dict of {str: list[str]}
+        Mapping from factor name or family name to the factor names or family names it
+        must be neutralized against. A family key neutralizes each factor in that family
+        independently against the same targets.
+
+    exposures : ndarray of shape (n_observations, n_assets, n_factors)
+        Factor exposures. Neutralized columns are overwritten in-place.
+
+    benchmark_weights : ndarray of shape (n_observations, n_assets)
+        Cross-sectional weights for neutralization and residual scaling.
+
+    factor_names : ndarray of shape (n_factors,)
+        Factor names.
+
+    factor_families : ndarray of shape (n_factors,)
+        Family label for each factor.
+
+    Raises
+    ------
+    ValueError
+        If a key or target is neither a factor name nor a family name, or if a key
+        resolves to factors that overlap with its neutralization targets.
+    """
+    factor_to_idx, family_to_idx = _factor_name_maps(factor_names, factor_families)
+
+    for key, targets in neutralize_against.items():
+        neutralize_idx = _resolve_factor_name(key, factor_to_idx, family_to_idx)
+        targets_list = _expand_factor_names(targets, factor_to_idx, family_to_idx)
+        targets_idx = set(targets_list)
+
+        overlap = neutralize_idx & targets_idx
+        if overlap:
+            overlap_names = sorted(factor_names[i] for i in overlap)
+            raise ValueError(
+                f"`neutralize_against` key '{key}' resolves to factors "
+                f"that overlap with its targets: {overlap_names}. "
+                f"A factor cannot be neutralized against itself."
+            )
+
+        x = exposures[:, :, targets_list]
+        for factor_idx in sorted(neutralize_idx):
+            neutralized, cs_weights = _cs_neutralize(
+                y=exposures[:, :, factor_idx],
+                x=x,
+                cs_weights=benchmark_weights,
+                cs_regressor=cs_regressor,
+            )
+            exposures[:, :, factor_idx] = CSStandardScaler().fit_transform(
+                neutralized, cs_weights=cs_weights
+            )
