+++ src/skfolio/utils/_factor_tools.py
@@ -10,9 +10,7 @@ from collections import defaultdict
 
 import numpy as np
 
-from skfolio.linear_model import BaseCSLinearModel, CSLinearRegression
-from skfolio.preprocessing import CSStandardScaler
-from skfolio.typing import FloatArray, ObjArray, StrArray
+from skfolio.typing import ObjArray, StrArray
 
 
 def _factor_name_maps(
@@ -204,219 +202,3 @@ def _resolve_factor_subset(
         if factor_family in family_names_to_keep
     ]
     return indices, [all_names[i] for i in indices]
-
-
-def _cross_sectional_neutralize(
-    y: FloatArray,
-    x: FloatArray,
-    cs_weights: FloatArray,
-    cs_regressor: BaseCSLinearModel | None = None,
-) -> tuple[FloatArray, FloatArray]:
-    r"""Neutralize a panel against cross-sectional explanatory variables.
-
-    For each observation :math:`t`, regress :math:`y_t` on :math:`X_t` across
-    assets with cross-sectional weights and return the residuals:
-
-    .. math::
-
-        \tilde{y}_t = y_t - X_t \hat{\beta}_t
-
-    Missing `y` values or rows of `x` with missing features are excluded by
-    setting their effective regression weights to zero.
-
-    Parameters
-    ----------
-    y : ndarray of shape (n_observations, n_assets)
-        Cross-sectional response values to neutralize.
-
-    x : ndarray of shape (n_observations, n_assets, n_features)
-        Cross-sectional explanatory variables.
-
-    cs_weights : ndarray of shape (n_observations, n_assets)
-        Base cross-sectional weights. Entries whose `y` or `x` values are
-        missing receive zero effective weight.
-
-    cs_regressor : BaseCSLinearModel, optional
-        Cross-sectional linear regressor used for residualization. If `None`,
-        `CSLinearRegression(fit_intercept=False)` is used.
-
-    Returns
-    -------
-    neutralized : ndarray of shape (n_observations, n_assets)
-        Cross-sectional residuals. Missing values in `y` or `x` propagate to
-        the corresponding residual entries.
-
-    weights : ndarray of shape (n_observations, n_assets)
-        Effective regression weights after missing-value exclusion.
-    """
-    valid = np.isfinite(y) & np.all(np.isfinite(x), axis=2)
-    cs_weights = np.where(valid, cs_weights, 0.0)
-    regressor = (
-        CSLinearRegression(fit_intercept=False)
-        if cs_regressor is None
-        else cs_regressor
-    )
-    neutralized = y - regressor.fit(x, y, cs_weights=cs_weights).predict(x)
-    return neutralized, cs_weights
-
-
-def _neutralize_scores(
-    neutralize_against: list[str],
-    scores: FloatArray,
-    exposures: FloatArray,
-    cs_weights: FloatArray,
-    factor_names: ObjArray,
-    factor_families: ObjArray | None = None,
-) -> FloatArray:
-    """Neutralize descriptor scores against selected factor exposures.
-
-    Parameters
-    ----------
-    neutralize_against : list of str
-        Factor names or family names to neutralize each descriptor score against.
-
-    scores : ndarray of shape (n_observations, n_assets, n_descriptors)
-        Descriptor score panels. The array is modified in-place.
-
-    exposures : ndarray of shape (n_observations, n_assets, n_factors)
-        Factor exposures used as neutralization variables.
-
-    cs_weights : ndarray of shape (n_observations, n_assets)
-        Cross-sectional weights for the neutralization regressions.
-
-    factor_names : ndarray of shape (n_factors,)
-        Factor names.
-
-    factor_families : ndarray of shape (n_factors,), optional
-        Family label for each factor. If provided, `neutralize_against` may contain
-        family names.
-
-    Returns
-    -------
-    scores : ndarray of shape (n_observations, n_assets, n_descriptors)
-        The input score array with each descriptor replaced by its neutralized residuals.
-
-    Raises
-    ------
-    ValueError
-        If a neutralization target is neither a factor name nor a family name.
-    """
-    factor_to_idx, family_to_idx = _factor_name_maps(factor_names, factor_families)
-    targets_idx = _expand_factor_names(neutralize_against, factor_to_idx, family_to_idx)
-    if len(targets_idx) == 0:
-        return scores
-
-    x = exposures[:, :, targets_idx]
-    for i in range(scores.shape[2]):
-        scores[:, :, i], _ = _cross_sectional_neutralize(
-            y=scores[:, :, i], x=x, cs_weights=cs_weights
-        )
-
-    return scores
-
-
-def _neutralize_exposures(
-    cs_regressor: BaseCSLinearModel,
-    neutralize_against: dict[str, list[str]],
-    exposures: FloatArray,
-    benchmark_weights: FloatArray,
-    factor_names: ObjArray,
-    factor_families: ObjArray,
-) -> None:
-    """Neutralize factor exposures against specified factors or families.
-
-    For each entry in `neutralize_against`, the key's exposures are regressed against
-    the target factors and replaced in-place by standardized residuals. Keys and targets
-    accept factor names or family names. Entries are processed in insertion order, so
-    later entries see exposures already modified by earlier entries.
-
-    Parameters
-    ----------
-    cs_regressor : BaseCSLinearModel
-        Cross-sectional linear regressor used for residualization.
-
-    neutralize_against : dict of {str: list[str]}
-        Mapping from factor name or family name to the factor names or family names it
-        must be neutralized against. A family key neutralizes each factor in that family
-        independently against the same targets.
-
-    exposures : ndarray of shape (n_observations, n_assets, n_factors)
-        Factor exposures. Neutralized columns are overwritten in-place.
-
-    benchmark_weights : ndarray of shape (n_observations, n_assets)
-        Cross-sectional weights for neutralization and residual scaling.
-
-    factor_names : ndarray of shape (n_factors,)
-        Factor names.
-
-    factor_families : ndarray of shape (n_factors,)
-        Family label for each factor.
-
-    Raises
-    ------
-    ValueError
-        If a key or target is neither a factor name nor a family name, or if a key
-        resolves to factors that overlap with its neutralization targets.
-    """
-    factor_to_idx, family_to_idx = _factor_name_maps(factor_names, factor_families)
-
-    for key, targets in neutralize_against.items():
-        neutralize_idx = _resolve_factor_name(key, factor_to_idx, family_to_idx)
-        targets_list = _expand_factor_names(targets, factor_to_idx, family_to_idx)
-        targets_idx = set(targets_list)
-
-        overlap = neutralize_idx & targets_idx
-        if overlap:
-            overlap_names = sorted(factor_names[i] for i in overlap)
-            raise ValueError(
-                f"`neutralize_against` key '{key}' resolves to factors "
-                f"that overlap with its targets: {overlap_names}. "
-                f"A factor cannot be neutralized against itself."
-            )
-
-        x = exposures[:, :, targets_list]
-        for factor_idx in sorted(neutralize_idx):
-            neutralized, cs_weights = _cross_sectional_neutralize(
-                y=exposures[:, :, factor_idx],
-                x=x,
-                cs_weights=benchmark_weights,
-                cs_regressor=cs_regressor,
-            )
-            exposures[:, :, factor_idx] = CSStandardScaler().fit_transform(
-                neutralized, cs_weights=cs_weights
-            )
-
-
-def _powered_market_cap(market_caps: FloatArray, power: float = 1.0) -> FloatArray:
-    r"""Compute :math:`\text{mcap}^{\text{power}}` as an unnormalized cap signal.
-
-    Returns the raw powered market capitalizations **without** normalizing
-    them to sum to one.  Downstream consumers (cross-sectional regression,
-    cross-sectional scalers, basis transforms, etc.) are responsible for
-    masking invalid entries and normalizing to their own validity context.
-
-    This follows the scikit-learn `sample_weight` convention: weights
-    express relative importance and need not sum to one.
-
-    Parameters
-    ----------
-    market_caps : array-like of shape (n_observations, n_assets)
-        Market capitalizations.  May contain NaN for delisted /
-        out-of-universe assets.
-
-    power : float, default=1.0
-        Exponent applied to market caps.  Typical choices:
-
-        * 0.0 -- equal weight (caller should handle separately)
-        * 0.5 -- sqrt-cap weight
-        * 1.0 -- cap weight (default)
-
-    Returns
-    -------
-    cap_signal : ndarray of shape (n_observations, n_assets)
-        :math:`\text{mcap}^{\text{power}}`.  NaN where `market_caps`
-        is NaN.
-    """
-    if power != 1.0:
-        market_caps = np.power(market_caps, power)
-    return market_caps
